internal rework
