+++ packages/orama/src/components/index.ts
@@ -29,7 +29,7 @@ import { RadixTree } from '../trees/radix.js'
 import { BKDTree } from '../trees/bkd.js'
 import { BoolNode } from '../trees/bool.js'
 
-import { convertDistanceToMeters } from '../utils.js'
+import { convertDistanceToMeters, setIntersection, setUnion } from '../utils.js'
 import { BM25 } from './algorithms.js'
 import { getMagnitude } from './cosine-similarity.js'
 import { getInnerType, getVectorSize, isArrayType, isVectorType } from './defaults.js'
@@ -823,94 +823,6 @@ export function createIndex(): IIndex<Index> {
     load,
     save
   }
-}
-
-
-const withIntersection = 'intersection' in (new Set());
-function setIntersection<V>(...sets: Set<V>[]): Set<V> {
-  // Fast path 1
-  if (sets.length === 0) {
-    return new Set();
-  }
-  // Fast path 2
-  if (sets.length === 1) {
-    return sets[0];
-  }
-  // Fast path 3
-  if (sets.length === 2) {
-    const set1 = sets[0];
-    const set2 = sets[1];
-
-    if (withIntersection) {
-      return set1.intersection(set2);
-    }
-    const result = new Set<V>();
-    const base = set1.size < set2.size ? set1 : set2;
-    const other = base === set1 ? set2 : set1;
-    for (const value of base) {
-      if (other.has(value)) {
-        result.add(value);
-      }
-    }
-    return result;
-  }
-
-  // Slow path
-  // Find the smallest set
-  const min = {
-    index: 0,
-    size: sets[0].size,
-  }
-  for (let i = 1; i < sets.length; i++) {
-    if (sets[i].size < min.size) {
-      min.index = i;
-      min.size = sets[i].size;
-    }
-  }
-
-  if (withIntersection) {
-    let base = sets[min.index];
-    for (let i = 0; i < sets.length; i++) {
-      if (i === min.index) {
-        continue;
-      }
-      base = base.intersection(sets[i]);
-    }
-
-    return base;
-  }
-
-  // manual implementation:
-  // intersect all sets with the smallest set
-  const base = sets[min.index];
-  for (let i = 0; i < sets.length; i++) {
-    if (i === min.index) {
-      continue;
-    }
-    const other = sets[i];
-    for (const value of base) {
-      if (!other.has(value)) {
-        base.delete(value);
-      }
-    }
-  }
-
-  return base;
-}
-
-const withUnion = 'union' in (new Set());
-function setUnion<V>(set1: Set<V> | undefined, set2: Set<V>) {
-  if (withUnion) {
-    if (set1) {
-      return set1.union(set2);
-    }
-    return set2;
-  }
-
-  if (!set1) {
-    return new Set(set2);
-  }
-  return new Set([...set1, ...set2]);
 }
 
 function addGeoResult(set: Set<InternalDocumentID> | undefined, ids: Array<{ docIDs: InternalDocumentID[] }>): Set<InternalDocumentID> {
