+++ packages/orama/src/utils.ts
@@ -341,3 +341,91 @@ export function isPromise(obj: any): obj is Promise<unknown> {
 export function isAsyncFunction(func: any): boolean {
   return func?.constructor?.name === 'AsyncFunction'
 }
+
+
+const withIntersection = 'intersection' in (new Set());
+export function setIntersection<V>(...sets: Set<V>[]): Set<V> {
+  // Fast path 1
+  if (sets.length === 0) {
+    return new Set();
+  }
+  // Fast path 2
+  if (sets.length === 1) {
+    return sets[0];
+  }
+  // Fast path 3
+  if (sets.length === 2) {
+    const set1 = sets[0];
+    const set2 = sets[1];
+
+    if (withIntersection) {
+      return set1.intersection(set2);
+    }
+    const result = new Set<V>();
+    const base = set1.size < set2.size ? set1 : set2;
+    const other = base === set1 ? set2 : set1;
+    for (const value of base) {
+      if (other.has(value)) {
+        result.add(value);
+      }
+    }
+    return result;
+  }
+
+  // Slow path
+  // Find the smallest set
+  const min = {
+    index: 0,
+    size: sets[0].size,
+  }
+  for (let i = 1; i < sets.length; i++) {
+    if (sets[i].size < min.size) {
+      min.index = i;
+      min.size = sets[i].size;
+    }
+  }
+
+  if (withIntersection) {
+    let base = sets[min.index];
+    for (let i = 0; i < sets.length; i++) {
+      if (i === min.index) {
+        continue;
+      }
+      base = base.intersection(sets[i]);
+    }
+
+    return base;
+  }
+
+  // manual implementation:
+  // intersect all sets with the smallest set
+  const base = sets[min.index];
+  for (let i = 0; i < sets.length; i++) {
+    if (i === min.index) {
+      continue;
+    }
+    const other = sets[i];
+    for (const value of base) {
+      if (!other.has(value)) {
+        base.delete(value);
+      }
+    }
+  }
+
+  return base;
+}
+
+const withUnion = 'union' in (new Set());
+export function setUnion<V>(set1: Set<V> | undefined, set2: Set<V>) {
+  if (withUnion) {
+    if (set1) {
+      return set1.union(set2);
+    }
+    return set2;
+  }
+
+  if (!set1) {
+    return new Set(set2);
+  }
+  return new Set([...set1, ...set2]);
+}
